projects/d05783bd-7ebe-456f-b716-00a75de1eace/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
sqlite3==3.38.0
pydantic==1.9.1
python-jose==3.4.0
